Bugfix: Don't spam the logger
When cron job is run frequently, "Cron run completed." Messages flood the logger.
This is a actually a rather significant problem because it will cause the quota to be reached more easily and important logs will be flushed out.
Drupal does not seem to provide a way to silence this without a manual patch as far as I can tell, so this is the only way to do this.