Bugfix: Append the error and required tags to the form-item block
This fixes the logic bug where fields do not have a way to specify error and required status via css.
With this change, the error and required status may be applied to the form-item object as a whole and not just the input field.
In particular, this allows for highlighting the label, input field, and description of a problematic field.
Many browsers do not allow changing the checkbox and radio buttons to
have a border of any kind, so doing this also allows for identifying problematic checkbox and radio button fields.