Feature: add support for uploading STREAM_WRAPPERS_REMOTE
STREAM_WRAPPERS_REMOTE are files stored remotely.
Normally, this can be done without a patch, but if files are to be uploaded to the remote address, then special handling is needed.

This is a requirement for the mcneese_file_db functionality.

Make sure to not replace 'public://' and 'private://' stream wrappers.

Add feature where one can specify the default image style scheme via the image_style_default_scheme variable.