Feature: allow for fieldsets to be replaced with div (required for printing)
There are different forms of fieldsets used in drupal theming.
1) directly created fieldsets.
2) indirectly created fieldset through other themes, namely radios and checkboxes.

Each case needs to be handled differently.
When printing, fieldsets do not generally print correctly (due to browser bugs) and must be converted to divs.
In some cases, such as with webforms, fieldsets provide the proper structure but the usage of fieldsets is unwanted.

This adds additional element keys:
- #as_div:
  This provides a way for fieldset themes to be manually created as a <div> instead of a <fieldset>.
  This is used by an appropriately patched webform module.
- #fieldset:
  This allows for radios and checkboxes to conditionally determine when to use a fieldset and when to use a div.
  This fixes compatibility problems with the way drupal core and other modules utilize radio buttons, checkboxes, radio, and checkbox (drupal core uses different themes for the plural forms and the singular forms).