Hack: Prevent drupal from checking updates when just viewing the update status page
I generally want to just view the current update status.
I do not want to be sitting on the page waiting for it to do a 5 minute update check whenever it feels like it.

Instead, if I want the latest results, then I will click the 'Check Manually' link.

This change removes the headache and allows for quickly viewing the update status without causing any update checks.