Hack: strip out invalid utf-8 characters
I noticed this problem when what is likely an attacker (or bot) was trying to find a way into breaking the web system.

The following is the URL they used, with example.com/some_page replacing the local real path:
http://example.com/some_page++++++++++++++++++++Result:+%ED%E5+%ED%E0%F8%EB%EE%F1%FC+%F4%EE%F0%EC%FB+%E4%EB%FF+%EE%F2%EF%F0%E0%E2%EA%E8;+Result:+%ED%E5+%ED%E0%F8%EB%EE%F1%FC+%F4%EE%F0%EC%FB+%E4%EB%FF+%EE%F2%EF%F0%E0%E2%EA%E8;

The $_GET['q'] has the url_decoded version. This decoded string is then passed to drupal_lookup_path() at some point in time.

This exposed two problems:
1) The SQL Queries inside of drupal_lookup_path() are not wrapped with try { ... } catch (Exception $ex) { ... } blocks.
- This results in WSOD on query errors.

2) MySQL and PostgreSQL handle invalid charset characters differently.
- MySQL appears to allow the invalid characters. (Is it dropped or ignored, and if it is ignored, then is there any potential security issue here?)
- PostgreSQL seems to throw an error: Character not in repertoire: 7 ERROR: invalid byte sequence for encoding "UTF8": 0xed 0xe5 0x20

To hackishly fix the invalid string, I tried the following inside of drupal_path_initialize():

$encoding = mb_detect_encoding($_GET['q'], "auto");
$valid_string = iconv($encoding, 'UTF-8//IGNORE', $_GET['q']);

The $encoding that was returned is UTF-8, but $valid_string evaluates to FALSE.

The hack that I came up with to get this to work, by stripping out the invalid characters is:

  $valid_string = iconv('UTF-8', 'UTF-8//IGNORE', $_GET['q']);
  if ($valid_string === FALSE) {
    $valid_string = iconv('ISO8859-1', 'UTF-8//IGNORE', $_GET['q']);
    if ($valid_string !== FALSE) $_GET['q'] = $valid_string;
  }

It appears that casting it to UTF-8 returns FALSE, then it is known that an invalid UTF-8 character exists.
The //IGNORE tells iconv() to strip out invalid characters and not return FALSE.

This is, unfortunately a hack, but it at least allows me to return a page not found screen, such as this simpler example:
http://example.com/h%ed

Before the hack, I would get: WSOD.
After the hack I would get: http://example.com/h%C3%AD
After the hack, using UTF-16, I would get:
http://example.com/%EE%B5%A8 (the h appears to have vanished).

Signed-off-by: Kevin Day <kday@mcneese.edu>