Hack: implement a hard-coded version of the 404 and 403 error pages
The original drupal design, while convenient, is built around logic flaws and design flaws.
This change prevents problems and allows for properly functioning custom error pages,

One of the flaws is that the 404 (page not found) does not process additional menus and module functions.
This results in parts of a page, such as the primary and secondary navigation, from appearing.
This behavior is very confusing and misleading.

Another flaw is that when custom pages are used, the users are redirect to those custom pages.
This causes the original path to not function and anything that should be appearing on an access denied path is prevented from appearing, such as a login block.

www-885