Hack: allow for streamwrappers to function without directories
This adds support for making sure certain functions don't fail when trying to create a directory for streamwrappers that do not support directories.
This is for the custom stream wrappers that don't care about directories, such as a database.

To utilize this function, the custom stream wrapper must define the following inside the class:
  static protected $directoryless = TRUE;