Update: don't use variable_get() to get local file paths
This should be done by just changing the file!
The big problem with the drupal 7 approach here is doing a database call to get a known local path.
Removing this allows for a performance improvement and also allows for the site to function better when the database is inaccessible.
(Which generally means throwing friendly errors to the screen instead of outright crashing all over the screen)

Signed-off-by: Kevin Day <kday@mcneese.edu>