Bugfix: remove extra processing code that changes LIKE to ILIKE
By default drupal changes the behavior of LIKE operations to ILIKE operations in postgresql to make it consistent with MySQLs violation of LIKE.
This is bad practice and causes postgresql to incur a performance penatly.

The change LIKE to ILIKE to make postgresql happy.

Note: After research, I learned in this case the SQL is either too vague or PostgreSQL is not following the standard.