Cleanup: remove update_init() to improve performance for users with 'administer site configuration' permissions
This shaves off about 100ms of load time to perform uncessary tasks on every admin page load for users with: 'administer site configuration'.

This function is more of a convenience.
A good admin will be able to check and know when their website is using insecure packages, so do not cost production runtime on something that can be checked separately at any point in time.

Signed-off-by: Kevin Day <kday@mcneese.edu>