Security: remove all files that do not need to be exposed publically, including those in robots.txt (revert this on new install)
All files in a post-installed system should not be exposed to the world.

It is important to note that for new systems, this commit must be reverted.

This decreases security by:
- helps prevent fingerprinting the system.
- reducing code and paths where an exploit may happen.
- helps reduce information that is exposed to the world.
- removes .htaccess, which should be disbled via apache and the file removed (if using .htaccess in apache, then this file must stay!).
- removes local scripts that should not be available on a web server front end.
- removes install.php so that it can never be executed on an already installed system.

A common exploit happens via execution of the cron.php, which is no longer possible because cron.php is now removed.