Security: remove all files that do not need to be exposed publicly, including those in robots.txt (revert this on new install)
All files in a post-installed system should not be exposed to the world.

It is important to note that for new systems, this commit must be reverted before install for proper installation and then re-applied after installation.

This increases security by:
- helps prevent fingerprinting the system to a specific version.
- reducing code and paths where an exploit may happen.
- helps reduce information that is exposed to the world.
- removes .htaccess, which should be disabled via apache and the file removed (if using .htaccess in apache, then this file must stay!).
- removes local scripts that should not be available on a web server front end.
- removes install.php so that it can never be executed on an already installed system.
- removes xmlrpc.php, thus avoiding constant xmlrpc security issues that always seem to crop up.
- removes cron.php, which avoids common exploits that tend to crop up.
- removes additional css, image, and javascript files from robots.txt (added as of drupal 7.50).
- removes profiles/minimal
- removes profiles/standard

Even with the profiles/standard and profiles/minimal directories back the warnings still happen.
- This is because the public front end uses symbolic links and is not setup to link to (and therefore expose) the profiles directory.

Remove the profile directory and then strip out the code that forces the profile code to exist.
Outside of "requirements" the profile provides no functional runtime benefits except for the libraries module usage.
The libraries module simply uses it to search for libraries (none of which are installed there).
- Therefore the libraries module can and should have this usage removed.