Security: remove authorize.php and athorize support
This functionality allows for changing files on the local systen through apache/php.
This is a major security issue as it allows any exploit on apache/php to arbitrarily alter the code on the system.

This is my first attempt at stripping this "functionality" out, keep an eye out for bugs.