Feature: optimize File Database load performance by avoiding drupal almost entirely
This is a custom php script that performs the necessary calls as the public_user account.
- No drupal code beyond the 'settings.php' and the 'mcneese_file_db.module' are accessed.

Performance considerations (time taken to receive image after requesting from server), using the file: 75th_anniversary-banner.png:
- Fully bootstrapped drupal: ~600 to ~800 ms.
- Minimally bootstrapped drupal: ~250 to ~300ms.
- No drupal code: ~29 to 35 ms.

This utilizes the custom error pages stored in /not_found/ .
This also supports the custom error page found at: /service_unavailable/ (which is not implemented at the time of this commit.).

This provides other error pages, but they are supplied with a simple white page and message.

To implement this on a site (example using wwwdev), edit apache inside the <Directory "/var/www/drupal/wwwdev"> section:
    RewriteRule ^f/c/ unrestricted.php [END]
    RewriteRule ^f/i/ unrestricted.php [END]
    RewriteRule ^f/f/ unrestricted.php [END]
    RewriteRule ^f/s/ unrestricted.php [END]

Make sure the above code is inside of the "<IfModule rewrite_module>", after "RewriteBase /" and before "RewriteRule ^ - [E=protossl]".

Note that [END] must be used to prevent the remaining RewriteRules from processing the unrestricted.php file.

Signed-off-by: Kevin Day <kday@mcneese.edu>