Bugfix: return value handling of database query for file_db in unrestricted.php not handling errors properly
The return value from a pg_query() call without a fetch command returns a databse object or FALSE.
The code seems to check if it is empty(), which is not valid.

Instead, call pg_fetch_array() on the result and then check if the return value (aka: $row) is empty() or is not an array.
This produces the expected behavior.