Hack: fix caching issues, add custom site-specific default header settings, fix generation of http headers
Fix problems with browser caching and allow users to navigated backwards and forwards via their browser navigation.
Cache is done in a private manner for logged in users and a public manner otherwise.
The login form explicitly is setup for private caching.

While working on this I noticed that because of the way the headers are generated and sent, some possible locations for header alterations get ignored.
This caused a few cases where headers were being altered to be ignored.
- The output must be cached completely and then both the headers and output must be sent.

New options are:
- default_headers-public = array of header fields and values.
- default_headers-private = array of header fields and values.
- default_headers-form = array of header fields and values.
- default_headers-last_modified = TRUE/FALSE
- default_headers-cache_control = TRUE/FALSE
- default_headers-expires = TRUE/FALSE

When default_headers-last_modified, default_headers-cache_control, default_headers-expires, are enabled, then defaults are assigned for those fields.

ticket: fcs-52
ticket: www-1148