Bugfix: "If you are already logged in.." is added multiple times
The error documents script already has the "If you are already logged in..", so if the drupal code appends it, then it appears twice.
Avoid this by leaving the "If you are already logged in.." text to the static error document pages.

Also add a <br> before the user login where applicable.