Bugfix: prevent divide by 0 in drupal core image_dimensions_scale()
The function does not check the variable when diving the height by the width.
If the width is not set or is less than 0, then set the scale to 1.0.
If the height is not set or less than 0, then set the scale to 1.0.

This issue is exposed when imagemagick fails to generate an image resize.

ticket: 1266