Bugfix: fix url encoding problems when generating file links
The function theme_file_link calls file_create_url() to generate a url.
Then the l() function is called to turn it into a link.

What happens is that rawurlencode() gets called by both file_create_url() and l().
This is a double-encoding bug.
The solution is to decode the url that is returned from file_create_url() before passing it to l().

Signed-off-by: Kevin Day <kday@mcneese.edu>