Security: revert drupals 'ajax security fix', it breaks functionality for a client-side securty bug
Client Side security is not relevant to the site as all ajax data should be treated and handled as untrusted already.

This so called 'fix' breaks behavior, does the opposite of what the code implies (in regards to the clean urls change), and exposes unecessary information to the client side that does nothing to improve security.
I consider their security fix a security issue.

According to drupal:
  "A vulnerability was found that allows a malicious user to perform a cross-site scripting attack by invoking Drupal.ajax() on a whitelisted HTML element."
  "This vulnerability is mitigated on sites that do not allow untrusted users to enter HTML."

see: https://www.drupal.org/SA-CORE-2015-003