Bugfix: Invalid use of Content-Length HTTP Header resulting from calling strlen()
RFC 7230 describes the content-length as referring to the count based on every 8-bits, which is a single octet.
Using strlen() would be incorrect because characters are 7-bit.
Using mb_strlen() would be incorrect because it contains mixed lengths.
Using any string test would be incorrect because the content may already be binary data.

The solution is to break apart the string into 8-bit chunks and then calculate the length.
This may be expensive depending on the size of the data and ideally should only be called once.

ticket: www-1315
ticket: ucs-1541