Feature: optimize /node/X paths (where X is a digit)
The /node/X paths generall are url redirects to a friendlier name.
The original behavior requires loading the entire drupal core information, which takes ~600ms in ideal situations.
Once the redirect happens, another ~600ms load must be performed to render the url.

This is incredibly inefficient.
Instead, provide a custom script (pathing.php) that will search the url_alias table in the database for a match for the /node/X urls.
If a match is found, then redirect to the discovered string.
This takes 90ms in ideal sitations.

When page loads take an unusually long take, say 2 seconds for example, that 2 seconds would become 4 seconds due to the original behavior.
This new behavior would take only 2 seconds (+90 milliseconds).

This is a major performance improvement.

The script is designed to fallback to the original drupal functions on failure.

The following rule in apache is required to utilize this:
  RewriteRule ^node/ pathing.php [END]

This supports PHP 7 Exception handling.