Feature: add postgresql-specific analyze command after cache clear all
It is recommended by postgres to perform an analyze immediately after major delete and update operations.
The cache clear all is one such example.

This only applies to the main cache tables.
No additional work is done on the other cache tables.
The other tables were avoided to keep this changeset simple.

Signed-off-by: Kevin Day <kday@mcneese.edu>