Feature: add support for postgresql-9.5+ UPSERT
There are problems with drupals caching process that results in database errors on cache_set().
By utilizing postgesql-9.5+'s upsert query support, this issue can hopefully be avoided.

The primary purpose of this is to solve a serious problem with drupals cache_set() function.
The functionality provided is only sufficient enough to perform the cache_set() operations.

The WHERE support for upserts is not implemented.

The cache operations need to be locked, so lock the cache table from updating until the operations are complete.
This helps prevent concurrency issues that continue to happen.

This will have a performance penalty in terms of updating the cache.
This should have a stability improvement as well as an performance improvement in terms of error logs writing to the disk.

see: https://www.postgresql.org/docs/current/static/sql-insert.html#SQL-ON-CONFLICT
see: https://www.postgresql.org/about/news/1636/
see: https://www.depesz.com/2015/05/10/waiting-for-9-5-add-support-for-insert-on-conflict-do-nothingupdate/
see: http://michael.otacoo.com/postgresql-2/postgres-9-5-feature-highlight-upsert/
see: http://stackoverflow.com/questions/17267417/how-to-upsert-merge-insert-on-duplicate-update-in-postgresql
see: https://www.compose.io/articles/postgresqls-future-is-looking-upsert/
see: http://pgeoghegan.blogspot.com/2015/10/avoid-naming-constraint-directly-when.html