Hack: do not allow 'Recoverable fatal errors' to be considered internal server errors
This one seems obvious.
If the errors can be recovered, then do not mark them as fatal!

This improves user experience by not providing an internal server error on problems.
The downside is that errors amy be printed to the screen.
The upside is that the errors are printed to the screen AND the site can still be navigated and used (depending on the specific error).