Hack: profile/standard and profile/minimal messages still being displayed, this time completely remove (revert this on new install)
Even with the profiles/standard and profiles/minimal directories back the warnings still happen.
- This is because the public front end uses symbolic links and is not setup to link to (and therefore expose) the profiles directory.

This time, try removing the profile directory and then strip out the code that forces the profile code to exist.
Outside of "requirements" the profile provides no functional runtime benefits except for the libraries module usage.
The libraries module simply uses it to search for libraries (none of which are installed there).
- Therefore the libraries module can and should have this usage removed.

The profiles/ directory removal code was removed from the commit "Security: remove all files that do not need to be exposed publicly, including those in robots.txt (revert this on new install)".
- This is done to keep these related changes self-contained.