Update: improve detection of and workaround problems when drupal_map_assoc() is mis-used
drupal_map_assoc() requires an array.
If an array is not passed, then PHP errors may occur.

This error is generally the result of a mis-use of the function (or possibly data is not validated/sanitized before being passed to drupal_map_assoc()).
Because this is a symptom of the real problem, there is nothing to resolve here.
Instead, additional verbosity is added to help detect, identify, and resolve the problem should it happen again.

I noticed that drupal also applies an additional check that is no longer necessary.

The following code now works as expected so there is no reason to work-around this other bug in drupal core.
  <?php
  var_dump (array_combine (array(), array()));
  ?>

ticket: www-1551
see: https://bugs.php.net/bug.php?id=34857
Signed-off-by: Kevin Day <kday@mcneese.edu>