Feature: provide a way to disable drupals 'query_string' add on via the variable 'drupal_disable_query_string'
This 'query_string' add on is applied to css and js files to try and 'force' the client browsers to change with cache clears.
This clobbers code, caching, and tries to force the will on the end-user (and is still not guaranteed to work).

Provide a way for site administers to optionally get rid of this mess.

Signed-off-by: Kevin Day <kday@mcneese.edu>